+++ lib/main.dart
@@ -1,9 +1,9 @@
-import 'dart:ui';
-
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 
 import '../gen/assets.gen.dart';
+import 'recently_added_list_item.dart';
+import 'recently_updated_list_item.dart';
 import 'themes.dart';
 
 void main() {
@@ -18,116 +18,6 @@ class MyApp extends StatelessWidget {
       title: 'Playlister',
       theme: darkTheme,
       home: const MyHomePage(title: 'Playlister'),
-    );
-  }
-}
-
-class RecentlyUpdatedListItem extends StatelessWidget {
-  final String text;
-  final Color color;
-
-  const RecentlyUpdatedListItem({
-    Key? key,
-    required this.text,
-    required this.color,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: 100,
-      margin: const EdgeInsets.symmetric(
-        horizontal: 10,
-      ),
-      child: InkWell(
-        onTap: () {},
-        borderRadius: BorderRadius.circular(
-          15,
-        ),
-        child: Ink(
-          decoration: BoxDecoration(
-            color: color,
-            border: Border.all(
-              color: Colors.white,
-            ),
-            borderRadius: BorderRadius.circular(
-              15,
-            ),
-          ),
-          child: Padding(
-            padding: const EdgeInsets.only(
-              left: 10,
-              right: 10,
-              bottom: 10,
-            ),
-            child: Align(
-              alignment: Alignment.bottomLeft,
-              child: Text(
-                text,
-              ),
-            ),
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-class RecentlyAddedListItem extends StatelessWidget {
-  final String text;
-  final Color color;
-
-  const RecentlyAddedListItem({
-    Key? key,
-    required this.text,
-    required this.color,
-  }) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      margin: const EdgeInsets.only(
-        top: 5,
-        bottom: 5,
-      ),
-      child: SizedBox(
-        height: 45,
-        child: InkWell(
-          onTap: () {},
-          borderRadius: BorderRadius.circular(
-            15,
-          ),
-          child: Ink(
-            decoration: BoxDecoration(
-              color: Colors.grey[800],
-              borderRadius: BorderRadius.circular(
-                15,
-              ),
-            ),
-            child: Row(
-              children: [
-                const SizedBox(
-                  width: 10,
-                ),
-                Container(
-                  decoration: BoxDecoration(
-                    color: color,
-                    shape: BoxShape.circle,
-                  ),
-                  height: 25,
-                  width: 25,
-                ),
-                const SizedBox(
-                  width: 10,
-                ),
-                Text(
-                  text,
-                ),
-              ],
-            ),
-          ),
-        ),
-      ),
     );
   }
 }
